Refactoring
- tabMoverPlugin became confusing, moved back to customBangs
